\name{SSLB-package}
\alias{SSLB-package}
\alias{SSLB}
\docType{package}
\title{
\packageTitle{SSLB}
}
\description{
\packageDescription{SSLB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SSLB}
\packageIndices{SSLB}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SSLB}

Maintainer: \packageMaintainer{SSLB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
