\name{SSLB-package}
\alias{SSLB-package}
\docType{package}
\title{
\packageTitle{SSLB}
}
\description{
\packageDescription{SSLB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SSLB}
\packageIndices{SSLB}
}
\author{
\packageAuthor{SSLB}

Maintainer: \packageMaintainer{SSLB}
}
\references{
Moran, G. E., Rockova, V. and George, E. I. (2020) "Spike-and-Slab Lasso Biclustering" Annals of Applied Statistics (Accepted)
}

